backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Mail
python-dotenv
PyJWT
Werkzeug
SQLAlchemy
python-dateutil
email-validator
bcrypt
gunicorn
